drawing ants as boxes and increase frame time
