+++ Cargo.toml
@@ -18,6 +18,7 @@ opt-level = 3
 rand = "0.8.5"
 getrandom = { version = "0.2", features = ["js"] }
 wasm-bindgen = "0.2.82"
+easybench-wasm = "0.2.1"
 
 [dependencies.web-sys]
 version = "0.3.4"
@@ -29,3 +30,4 @@ features = [
   'Window',
   'console'
 ]
+
