+++ index.js
@@ -1,4 +1,3 @@
 import('./pkg').then(wasm => {
   wasm.run();
 }).catch(console.error);
-
