--- src/ant.rs
@@ -1,6 +1,0 @@
-#[derive(Clone, Debug)]
-pub struct Ant {
-    pub x: f64,
-    pub y: f64,
-    pub dir: f64,
-}
