+++ src/consts.rs
@@ -1,10 +1,10 @@
 pub const PI: f64 = std::f64::consts::PI;
 
 pub const WANDER_COEFFICIENT: f64 = 0.3;
-pub const WALK_SPEED: f64 = 0.5;
-pub const ANT_RADIUS: f64 = 2.0;
-pub const ANT_COUNT: usize = 1000;
+pub const WALK_SPEED: f64 = 1.0;
+pub const ANT_RADIUS: f64 = 1.0;
+pub const ANT_COUNT: usize = 6000;
 
 pub const NEST_RADIUS: f64 = 10.0;
 
-pub const FRAME_TIME: i32 = 5;
+pub const FRAME_TIME: i32 = 17;
