--- src/helpers.rs
@@ -1,5 +1,0 @@
-use wasm_bindgen::JsValue;
-
-pub fn log(val: &str) {
-    web_sys::console::log_1(&JsValue::from_str(val));
-}
